requirements.txt
openai>=0.28.1
flask
Flask
flask_caching
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1